fix(grid): uint32 overflow in index calculation of large hashgrids
This fix affects grid sizes that mostly aren't used in practice. On the
order of 2^16 and up, but certain configurations might trigger the bug
for smaller grids as well. The bug is not critical in that it does not
lead to complete failure but merely to difficulty training some of the
levels of the grid.